separated the node from the interface and move to package model.service, I decided don't use Builder pattern, because object very simple.
