+++ src/main/java/executor/service/model/service/ContinuousJobNode.java
@@ -1,0 +1,24 @@
+package executor.service.model.service;
+
+import executor.service.service.parallelflowexecutor.Jobable;
+
+public class ContinuousJobNode {
+    private final Jobable jobable;
+    private Thread thread;
+
+    public ContinuousJobNode(Jobable jobable) {
+        this.jobable = jobable;
+    }
+
+    public Jobable getTask() {
+        return jobable;
+    }
+
+    public Thread getThread() {
+        return thread;
+    }
+
+    public void setThread(Thread thread) {
+        this.thread = thread;
+    }
+}
