+++ src/main/java/executor/service/service/parallelflowexecutor/ContinuousJobOperator.java
@@ -1,5 +1,7 @@
 package executor.service.service.parallelflowexecutor;
 
+import executor.service.model.service.ContinuousJobNode;
+
 import java.util.List;
 
 public interface ContinuousJobOperator {
@@ -8,26 +10,4 @@ public interface ContinuousJobOperator {
 
     void keepTaskAlive();
 
-    //TODO вынести в model и создании через Builder
-    class ContinuousJobNode {
-
-        private final Jobable jobable;
-        private Thread thread;
-
-        public ContinuousJobNode(Jobable jobable) {
-            this.jobable = jobable;
-        }
-
-        public Jobable getTask() {
-            return jobable;
-        }
-
-        public Thread getThread() {
-            return thread;
-        }
-
-        public void setThread(Thread thread) {
-            this.thread = thread;
-        }
-    }
 }
